Update submitjob.py with test function

+++ SampleMgr/python/18Selection.py
@@ -29,7 +29,7 @@ process.Jet = cms.PSet(
         
         AbsEta = cms.double( 2.4 ),
         Pt = cms.double( 30 ),
-        Pt_CRDYJets = cms.double( 60 ),
+        Pt_CRDYJets = cms.double( 30 ),
         Pre_AbsEta = cms.double( 2.4  ),
         Pre_Pt = cms.double( 20 ),
